+++ main.js
@@ -619,8 +619,10 @@ function fastLoop(){
 
         // trade routes
         if (global.tech['trade']){
+            let used_trade = 0;
             Object.keys(global.resource).forEach(function (res){
                 if (global.resource[res].trade > 0){
+                    used_trade += global.resource[res].trade;
                     let price = tradeBuyPrice(res) * global.resource[res].trade;
 
                     if (global.resource.Money.amount >= price * time_multiplier){
@@ -632,6 +634,7 @@ function fastLoop(){
                     steelCheck();
                 }
                 else if (global.resource[res].trade < 0){
+                    used_trade -= global.resource[res].trade;
                     let price = tradeSellPrice(res) * global.resource[res].trade;
 
                     if (global.resource[res].amount >= time_multiplier){
@@ -643,6 +646,7 @@ function fastLoop(){
                     steelCheck();
                 }
             });
+            global.city.market.trade = used_trade;
         }
 
         let power_grid = 0;
@@ -2907,10 +2911,13 @@ function midLoop(){
             global.civic.farmer.max = 0;
         }
 
-        if (global.race['kindling_kindred'] && global.civic.lumberjack.workers > 0){
+        if (global.race['kindling_kindred']){
             global.civic.lumberjack.workers = 0;
+            global.resource.Lumber.crates = 0;
+            global.resource.Lumber.containers = 0;
+            global.resource.Lumber.trade = 0;
         }
-        if (global.race['kindling_kindred'] && global.city.foundry['Plywood'] > 0){
+        if (global.race['kindling_kindred'] && global.city['foundry'] && global.city.foundry['Plywood']){
             global.city.foundry['Plywood'] = 0;
         }