+++ changelog.md
@@ -1,4 +1,5 @@
 #### 0.4.32 - 7/21/2019 ####
+Balorg slaver trait added<br>
 Ziggurat bonus now applies to the oil extractor<br>
 Geology Oil rich/poor no longer applies to oil extractor<br>
 Added resource alternate row coloring<br>