+++ changelog.md
@@ -1,3 +1,6 @@
+#### 0.4.33 - 7/22/2019 ####
+Fixes for Ent Fanatiiscm not releasing some resources<br>
+
 #### 0.4.32 - 7/21/2019 ####
 Balorg slaver trait added<br>
 Ziggurat bonus now applies to the oil extractor<br>